Finish async summary handler + simplify
